Ok - big internal change which separates the graphical representation from the model itself. Works fine.
Related to this - ModelItem is now an interface (not sure why it wasn't to start with, but there you go...)

Method names are lame - sue me (I need to alter a couple).

Graphical Model needs fixing in its 'objectATPoint' - it's doing it in the wrong order at the moment.

Should make Swing (or other) ports easier.